bot/modules/language: convert to TS
Convert bot/modules/language to TypeScript.
