--- bot/modules/language/index.js
@@ -1,8 +1,0 @@
-const { userMiddleware } = require('../../middleware/user');
-const commands = require('./commands');
-const actions = require('./actions');
-
-exports.configure = bot => {
-  bot.command('setlang', userMiddleware, commands.setlang);
-  bot.action(/^setLanguage_([a-z]{2})$/, userMiddleware, actions.setLanguage);
-};
